fix: set up winter-utils as a no-std crate
This commit tweaks winter-utils so that rather than being a std-using
crate that conditionally becomes a no-std crate, it becomes a no-std
crate which conditionally adds functionality from libstd when the `std`
feature is enabled

NOTE: This makes the `boxed`, `collections`, and `string` modules
redundant (they technically were before too, but due to how the crate
was being built, it wasn't obvious). A subsequent commit will mark those
deprecated, and update all downstream crates to use stuff from liballoc
or libstd directly (the latter for crates which are never built no-std).