fix(utils): incorrect buffer length check in slice reader impl
When determining whether we have sufficient bytes available to read a
slice, we were incorrectly looking at the total length of the current
buffer, not the portion of the buffer that is unread. As a result, we
wouldn't attempt to fill the buffer with more bytes, and a subsequent
attempt to read the requisite bytes from the buffer would fail, because
there weren't in fact enough to fufill the request from the unread
portion of the buffer.